Add SearchController in MatchmakingService
